complete?
Still needs testing across a multitude of chains, but pushing this commit so as to not lose track of the progress made here.

Have added support for essentially end to end vault creation with not much more than a deposit txn. This also adds functionality to pull pricing from CoinGecko's API for pricing, but as there are some gaps in CG's coverage, some manual input may be required.